refactor(factories): consolidate factory files from 14 to 8
Merge factory files to simplify navigation and reduce duplication:

Consolidations:
- data_source_registry.py + data_sources.py → data_source_factory.py
- base_services_factory.py + services_builder.py + runner_services.py → services_factory.py
- generic_factory.py + runner_assembly.py → pipeline_factory.py
- storage_factory.py + storage_adapter.py → storage.py

Deletions:
- clients.py (empty, deprecated)

Kept as-is:
- pipeline_factories.py (factory instances)
- http_client_factory.py
- transformer_factory.py
- __init__.py

Updated:
- All imports across codebase
- Test patches to use new module paths
- Architecture test exemptions for consolidated files

Result: 14 files → 8 files (≤8 requirement met)
