+++ tests/unit/composition/test_registry_protocol.py
@@ -74,7 +74,7 @@ class TestDataSourceRegistryUnifiedAPI:
     @pytest.fixture(autouse=True)
     def setup_and_teardown(self):
         """Restore registry after test."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -87,7 +87,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_list_keys_returns_list(self):
         """DataSourceRegistry.list_keys() should return a list."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -96,7 +96,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_list_keys_matches_list_providers(self):
         """list_keys() should return same result as list_providers()."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -104,7 +104,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_contains_returns_true_for_registered(self):
         """contains() should return True for registered providers."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -112,7 +112,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_contains_returns_false_for_unknown(self):
         """contains() should return False for unknown provider."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -120,7 +120,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_clear_empties_registry(self):
         """clear() should empty the registry."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -136,7 +136,7 @@ class TestDataSourceRegistryUnifiedAPI:
 
     def test_get_raises_keyerror_for_unknown(self):
         """get() should raise KeyError for unknown provider."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -156,7 +156,7 @@ class TestUnifiedAPIConsistency:
 
     def test_both_registries_have_list_keys(self):
         """Both registries should have list_keys() method."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -167,7 +167,7 @@ class TestUnifiedAPIConsistency:
 
     def test_both_registries_have_contains(self):
         """Both registries should have contains() method."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -178,7 +178,7 @@ class TestUnifiedAPIConsistency:
 
     def test_both_registries_have_clear(self):
         """Both registries should have clear() method."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -189,7 +189,7 @@ class TestUnifiedAPIConsistency:
 
     def test_both_registries_have_get(self):
         """Both registries should have get() method."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
 
@@ -200,7 +200,7 @@ class TestUnifiedAPIConsistency:
 
     def test_both_registries_have_register(self):
         """Both registries should have register() method."""
-        from bioetl.composition.factories.data_source_registry import (
+        from bioetl.composition.factories.data_source_factory import (
             DataSourceRegistry,
         )
