+++ tests/unit/infrastructure/factories/test_factories.py
@@ -40,7 +40,7 @@ class TestStorageAdapter:
     @pytest.fixture
     def storage_adapter(self, mock_bronze_writer, mock_silver_writer, mock_gold_writer):
         """Create StorageAdapter instance."""
-        from bioetl.composition.factories.storage_factory import StorageAdapter
+        from bioetl.composition.factories.storage import StorageAdapter
 
         return StorageAdapter(
             bronze_writer=mock_bronze_writer,
