+++ tests/unit/infrastructure/factories/test_storage_adapter.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 
 import pytest
 
-from bioetl.composition.factories.storage_factory import StorageAdapter
+from bioetl.composition.factories.storage import StorageAdapter
 from bioetl.domain.ports import StoragePort
 from bioetl.domain.types import RunType
